Refactor: Decompose PR review service into modular components for improved maintainability and readability. Key modules include file processing, diff analysis, AI analysis, and GitHub integration, enhancing separation of concerns and testability.
